+++ core/events/events_comment_change.ts
@@ -12,7 +12,7 @@
 // Former goog.module ID: Blockly.Events.CommentChange
 
 import * as registry from '../registry.js';
-import type {WorkspaceComment} from '../workspace_comment.js';
+import type {WorkspaceComment} from '../comments/workspace_comment.js';
 
 import {CommentBase, CommentBaseJson} from './events_comment_base.js';
 import * as eventUtils from './utils.js';
