+++ tests/mocha/event_comment_create_test.js
@@ -21,12 +21,9 @@ suite('Comment Create Event', function () {
 
   suite('Serialization', function () {
     test('events round-trip through JSON', function () {
-      const comment = new Blockly.WorkspaceComment(
-        this.workspace,
-        'test text',
-        10,
-        10,
-      );
+      const comment = new Blockly.comments.WorkspaceComment(this.workspace);
+      comment.setText('test text');
+      comment.moveTo(new Blockly.utils.Coordinate(10, 10));
       const origEvent = new Blockly.Events.CommentCreate(comment);
 
       const json = origEvent.toJson();
