+++ tests/mocha/event_comment_move_test.js
@@ -21,14 +21,11 @@ suite('Comment Move Event', function () {
 
   suite('Serialization', function () {
     test('events round-trip through JSON', function () {
-      const comment = new Blockly.WorkspaceComment(
-        this.workspace,
-        'test text',
-        10,
-        10,
-      );
+      const comment = new Blockly.comments.WorkspaceComment(this.workspace);
+      comment.setText('test text');
+      comment.moveTo(new Blockly.utils.Coordinate(10, 10));
       const origEvent = new Blockly.Events.CommentMove(comment);
-      comment.moveBy(10, 10);
+      comment.moveTo(new Blockly.utils.Coordinate(20, 20));
       origEvent.recordNew();
 
       const json = origEvent.toJson();
