+++ tests/mocha/index.html
@@ -121,6 +121,7 @@
       import './variable_model_test.js';
       import './blocks/variables_test.js';
       import './widget_div_test.js';
+      import './comment_view_test.js';
       import './workspace_comment_test.js';
       import './workspace_svg_test.js';
       import './workspace_test.js';
